Refactor partido creation and validation logic to ensure teams exist and are not the same

+++ routers/Equipos.py
@@ -1,7 +1,6 @@
 from typing import List
 from fastapi import APIRouter, HTTPException
 from modelos.Equipos import Equipo, EquipoCrear, EquipoActualizar
-from utils.verificaciones import *
 from db import SessionDep
 
 router = APIRouter(prefix="/equipos", tags=["equipos"])
@@ -21,15 +20,15 @@ async def read_equipos(session: SessionDep):
 @router.get("/{equipo_id}", response_model=Equipo)
 async def read_equipo(equipo_id: int, session: SessionDep):
     equipo = session.get(Equipo, equipo_id)
-    validar_equipo_existe(Equipo.id_equipo, session, "equipo")
+    if not equipo:
+        raise HTTPException(status_code=404, detail="El equipo no existe")
     return equipo
 
 @router.delete("/{equipo_id}", response_model=Equipo)
 async def delete_equipo(equipo_id: int, session: SessionDep):
     equipo = session.get(Equipo, equipo_id)
-
-    validar_equipo_existe(Equipo.id_equipo, session, "equipo")
-
+    if not equipo:
+        raise HTTPException(status_code=404, detail="El equipo no existe")
     session.delete(equipo)
     session.commit()
     return equipo
@@ -37,7 +36,8 @@ async def delete_equipo(equipo_id: int, session: SessionDep):
 @router.patch("/{equipo_id}", response_model=Equipo)
 async def update_equipo(equipo_id: int, updated_equipo: EquipoActualizar, session: SessionDep):
     equipo = session.get(Equipo, equipo_id)
-    validar_equipo_existe(Equipo.id_equipo, session, "equipo")
+    if not equipo:
+        raise HTTPException(status_code=404, detail="El equipo no existe")
     equipo_data = updated_equipo.model_dump(exclude_unset=True)
     for key, value in equipo_data.items():
         setattr(equipo, key, value)
